requirements.txt
Flask
flask-cors
scipy
peakutils
numpy
panda
psycopg2-binary
firebase-admin
pyrebase
sklearn